// nextjs-frontend/src/model/RegistrationForm.ts
export interface RegistrationForm {
    companyName: string;
    vatNumber: string;
    companyWebsite: string;
    productCategory: string[];
    streetAddress: string;
    streetNumber: string;
    city: string;
    zipCode: string;
    country: string;
    firstName: string;
    lastName: string;
    emailAddress: string;
    password: string;
}